Update README.md
Due to problems with commits I changed how I had started with my comments and explanations. Unsure if Laptop has an issue. I will not be amending code from this point but will just try to expand the read me the best I can. This is the reason for many of  the commits on readme.